Generalize workaround for floating point errors
The past workaround for floating point errors was to not accept
intersections which generated a very short distance.
The workaround had to be implemented for every object and it would
have been easy to forget it for future objects.

The new approach will move the point of intersection a tiny distance
along the surface normal and will work for any type of object.

The new approach will also fix some nasty anti-aliasing bugs...

This commit reverses 830d62